+++ config/dynamic.js
@@ -102,5 +102,7 @@ module.exports = {
 		memcachedServers,
 		port: 8888,
 		sessionUri: `https://www.${baseUrl}/start-ui-session`,
+		minVueWorkers: parseInt(process.env.MIN_VUE_WORKERS) || 1,
+		maxVueWorkers: parseInt(process.env.MAX_VUE_WORKERS) || 3,
 	}
 }