+++ server/dev-server.js
@@ -14,13 +14,14 @@ const webpack = require('webpack');
 const webpackDevMiddleware = require('webpack-dev-middleware');
 const webpackHotMiddleware = require('webpack-hot-middleware');
 const threadLoader = require('thread-loader');
-const promBundle = require("express-prom-bundle");
+const promBundle = require('express-prom-bundle');
+
 const metricsMiddleware = promBundle({
-    includeMethod: true,
-    includePath: true,
-    includeStatusCode: true,
-    includeUp: true,
-    promClient: {
+	includeMethod: true,
+	includePath: true,
+	includeStatusCode: true,
+	includeUp: true,
+	promClient: {
 		collectDefaultMetrics: {}
 	}
 });
@@ -40,7 +41,7 @@ const config = require('../config/selectConfig')(argv.config || 'local');
 const initCache = require('./util/initCache');
 const logger = require('./util/errorLogger');
 
-// Initialize a Cache instance, Should Only be called once!
+// Initialize a Cache instance
 const cache = initCache(config.server);
 
 // app init
@@ -112,7 +113,6 @@ const updateHandler = () => {
 			clientManifest,
 			serverBundle,
 			config,
-			cache,
 		});
 		resolveHandlerReady();
 	}
@@ -162,7 +162,7 @@ app.use(locale(config.app.locale.supported, config.app.locale.default));
 app.use('/ui-routes', serverRoutes);
 
 // Apply sitemap middleware to expose routes we want search engine crawlers to see
-app.use('/sitemaps/ui.xml', sitemapMiddleware(config.app, cache));
+app.use('/sitemaps/ui.xml', sitemapMiddleware(config.app, config.server));
 
 // Handle time sychronization requests
 app.use('/', timesyncRouter());
@@ -175,7 +175,7 @@ app.use(devMiddleware);
 app.use(hotMiddleware);
 
 // load metrics middleware
-app.use(metricsMiddleware)
+app.use(metricsMiddleware);
 
 // Configure session
 app.use('/', sessionRouter(config.server));